+++ react-data-grid/react-data-grid-tests.tsx
@@ -1,6 +1,6 @@
-/// <reference path="./react-data-grid.d.ts" />
-/// <reference path="../react/react.d.ts" />
-/// <reference path="../faker/faker.d.ts" />
+/// <reference types="react-data-grid" />
+/// <reference types="react" />
+/// <reference types="faker" />
 
 import * as React from 'react';
 import * as ReactDataGrid from 'react-data-grid';
@@ -11,8 +11,6 @@ var Editors = ReactDataGridPlugins.Editors;
 var Toolbar = ReactDataGridPlugins.Toolbar;
 var AutoCompleteEditor = Editors.AutoComplete;
 var DropDownEditor = Editors.DropDownEditor;
-
-faker.locale = 'en_GB';
 
 function createFakeRowObjectData(index:number):Object {
     return {
@@ -202,7 +200,7 @@ class Example extends React.Component<any, any> {
     getColumns() {
         var clonedColumns = columns.slice();
         clonedColumns[2].events = {
-            onClick: function (ev:React.SyntheticEvent, args:{idx:number, rowIdx:number}) {
+            onClick: function (ev:React.SyntheticEvent<{}>, args:{idx:number, rowIdx:number}) {
                 var idx = args.idx;
                 var rowIdx = args.rowIdx;
                 this.refs.grid.openCellEditor(rowIdx, idx);
